yaml/json conversion to separate module
