+++ settings.gradle.kts
@@ -54,6 +54,7 @@ include("serialization-core")
 include("serialization-json")
 include("serialization-xml")
 include("serialization-yaml")
+include("serialization-yaml-json-conversion")
 include("settings")
 include("value-enumeration")
 include("value-identifier")
